lara/col/climb: separate climb routines
This introduces collusion routine registration and initially splits
the climbing routines into their own module.
