+++ src/libtrx/meson.build
@@ -169,6 +169,7 @@ sources = [
   'game/lara/cheat.c',
   'game/lara/cheat_keys.c',
   'game/lara/col.c',
+  'game/lara/col/climb.c',
   'game/lara/common.c',
   'game/lara/control.c',
   'game/lara/flare.c',
